# optimized key for LRU hash (2x perfomance)