@SamSaffron 2020-07-21
We can not use hash, that would mean we can get collisions, in some cases you would look up an item from the cache and get a wrong materializer which would mean you would have to validate you got what you asked for.