@ermolaev 2020-07-21
`@cache` have size 500 items, collisions are unlikely

https://ruby-doc.org/core-2.7.0/Array.html#method-i-hash

![image](https://user-images.githubusercontent.com/938305/88034042-c9af1580-cb48-11ea-81df-522bd7b60b7d.png)

https://github.com/ruby/ruby/blob/1020e120e060c00eca456d5a129c344daa472407/hash.c#L194