@SamSaffron 2020-07-22
Sorry but these are the absolute worst type of production bugs to get. 3 times a year DB.query("something") does not work correctly due to a hash collision. 

I am afraid I can not merge this.